+++ pet-clinic-data/src/main/java/org/learning/spring/guruspringpetclinic/services/SpecialtyService.java
@@ -4,5 +4,5 @@ import org.learning.spring.guruspringpetclinic.model.Speciality;
 import org.springframework.stereotype.Service;
 
 @Service
-public interface SpecialtyService extends CrudService<Speciality, Long> {
+public interface SpecialtyService extends CrudService<Speciality, Integer> {
 }
