+++ pet-clinic-data/src/main/java/org/learning/spring/guruspringpetclinic/services/map/PetTypeServiceMap.java
@@ -5,5 +5,5 @@ import org.learning.spring.guruspringpetclinic.services.PetTypeService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PetTypeServiceMap extends AbstractMapService<PetType, Long> implements PetTypeService {
+public class PetTypeServiceMap extends AbstractMapService<PetType, Short> implements PetTypeService {
 }
